Move redirection spec helpers into a module
[#89128328]
