+++ spec/helpers/redirection.rb
@@ -1,0 +1,21 @@
+module Bookbinder
+  module Redirection
+    def capture_stdout(&block)
+      real_stdout = $stdout
+      $stdout = StringIO.new
+      block.call
+      $stdout.rewind
+      $stdout.read
+    ensure
+      $stdout = real_stdout
+    end
+
+    def swallow_stdout(&block)
+      real_stdout = $stdout
+      $stdout = StringIO.new
+      block.call
+    ensure
+      $stdout = real_stdout
+    end
+  end
+end
